rtl: don't lock for memcpy
Unlocks _buf_lock while data is copied to the sample buffer queue tail,
to allow the work function to proceed during this write.

As the write is to the tail, and the read is from the head, and changes
to the pointers are guarded by locks, no lock for this write is needed.